day5/pyproject.toml
#configuring toml
[project]
name = "bankingsystem"
version = "1.0.0"
description = "bankingsystem"
readme = "README.md"
authors = [
    {name = "bhavani" , email = "bhavanichittawadagi76@gmail.com"}
]
requires-python = ">=3.8"
#this is production dependencies
dependencies = [
    "pydantic>=2.0.0",
    "python-dotenv>=0.19.0",
    "faker>=10.0.0",
    "pandas>=2.0.0",
    "pydantic[email]>=2.0.0"
]
#development dependencies
[project.optional-dependencies]
dev = [
    "pytest>=7.2.0",
    "pytest-html>=4.2.0",
    "pytest-cov>=3.0.0",
    "black>=23.0.0",
    "isort>=5.10.0",
    "sphinx>=5.0.0"
]
[tool.black]
line-length = 88
target-version = ['py313']

[tool.setuptools]

package-dir = {"bankingsystem" = "."}

[tool.setuptools.packages.find]

where = ["."]